move collaborative_scaling to solidago
