+++ solidago/tests/test_collaborative_scaling.py
@@ -3,6 +3,7 @@ import pandas as pd
 
 from solidago.collaborative_scaling import estimate_positive_score_shift
 
+
 def test_score_shift_when_all_scores_are_equal():
     scaled_individual_scores = pd.DataFrame(dict(
         user_id=[f"user_{i}" for i in range(1000)],
